Move print_unix and print_tcptpi to src
